# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "QuantumClaw API",
    "description": "AI API Gateway & Token Distribution Platform",
    "version": "1.0.0",
    "contact": { "email": "support@quantumclaw.com" }
  },
  "host": "localhost:3666",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer token authentication. Format: Bearer <your_token>"
    }
  },
  "paths": {
...
    "/api/log": {
      "get": {
        "tags": ["Log"],
        "summary": "Query request logs",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "model", "in": "query", "type": "string"},
          {"name": "startIdx", "in": "query", "type": "integer", "default": 0},
          {"name": "num", "in": "query", "type": "integer", "default": 10}
        ],
        "responses": { "200": { "description": "Log list" } }
      }
    },
...
    "/api/cascade/node/register": {
      "post": {
        "tags": ["Cascade"],
        "summary": "Register a new cascade slave node",
        "description": "Used by slave nodes on startup to register with the master",
        "parameters": [{"name": "body", "in": "body", "required": true, "schema": {
          "type": "object",
          "properties": {
            "name": {"type": "string"},
            "region": {"type": "string"},
            "version": {"type": "string"}
          }
        }}],
        "responses": { "200": { "description": "Node ID + API Key" } }
      }
    },
...
    "/v1/models": {
      "get": {
        "tags": ["Relay"],
        "summary": "List available models (OpenAI-compatible)",
        "security": [{"BearerAuth": []}],
        "responses": { "200": { "description": "Model list" } }
      }
    },
...
  },
...
}